+++ packages/sitecore-jss-nextjs/src/components/NextImage.tsx
@@ -13,7 +13,7 @@ import Image, { ImageProps as NextImageProperties } from 'next/image';
 type NextImageProps = ImageProps & Partial<NextImageProperties>;
 
 export const NextImage: React.FC<NextImageProps> = withFieldMetadata<NextImageProps>(
-  ({ editable, imageParams, field, mediaUrlPrefix, fill, priority, ...otherProps }) => {
+  ({ editable = true, imageParams, field, mediaUrlPrefix, fill, priority, ...otherProps }) => {
     // next handles src and we use a custom loader,
     // throw error if these are present
     if (otherProps.src) {
@@ -97,10 +97,6 @@ NextImage.propTypes = {
   imageParams: PropTypes.objectOf(
     PropTypes.oneOfType([PropTypes.number.isRequired, PropTypes.string.isRequired]).isRequired
   ),
-};
-
-NextImage.defaultProps = {
-  editable: true,
 };
 
 NextImage.displayName = 'NextImage';